added fixture to api
