+++ crates/holochain_persistence_api/src/eav/eavi.rs
@@ -269,44 +269,6 @@ impl ExampleEntry {
     }
 }
 
-pub fn test_entry_a() -> ExampleEntry {
-    ExampleEntry::new(String::from("a"))
-}
-
-pub fn test_entry_b() -> ExampleEntry {
-    ExampleEntry::new(String::from("b"))
-}
-
-pub fn test_eav_entity() -> ExampleEntry {
-    test_entry_a()
-}
-
-pub fn test_eav_attribute() -> ExampleAttribute {
-    ExampleAttribute::WithPayload("foo-attribute".into())
-}
-
-pub fn test_eav_value() -> ExampleEntry {
-    test_entry_b()
-}
-
-pub fn test_eav() -> EntityAttributeValueIndex<ExampleAttribute> {
-    EntityAttributeValueIndex::new_with_index(
-        &test_eav_entity().address(),
-        &test_eav_attribute(),
-        &test_eav_value().address(),
-        0,
-    )
-    .expect("Could not create eav")
-}
-
-pub fn test_eav_content() -> Content {
-    test_eav().content()
-}
-
-pub fn test_eav_address() -> Address {
-    test_eav().address()
-}
-
 pub fn eav_round_trip_test_runner<A: Attribute>(
     entity_content: impl AddressableContent + Clone,
     attribute: A,
@@ -382,6 +344,10 @@ pub fn eav_round_trip_test_runner<A: Attribute>(
 #[cfg(test)]
 pub mod tests {
     use super::*;
+    use fixture::test_eav_entity;
+    use fixture::test_eav;
+    use fixture::test_eav_address;
+    use fixture::test_eav_content;
     use crate::{
         cas::{
             content::{AddressableContent, AddressableContentTestSuite, ExampleAddressableContent},
